requirements_consumer.txt
pika==1.2.0
python_dotenv==0.19.2
flask==1.1.2
itsdangerous==1.1.0
requests==2.22.0
